Sprint 6 findings: llama.cpp control-vector pipeline works, dataset needs rework
CONFIRMADO funcionando:
- extract_vector en modelo abliterado (Evil Ganda) via MPS, ~10min
- llama-completion (NO llama-cli, que es REPL) aplica control vectors one-shot
- Vector GGUF bien formado: 27 direction.N tensors, hidden 14336
- scripts/run_steered.sh para inference con --control-vector-scaled FNAME:SCALE

PROBLEMA identificado:
- Vector tiene efecto pero requiere alpha ~20 (magnitud repeng ~0.01)
- A alpha 20-40 rompe coherencia (loops) en vez de dar 'miedo'
- Causa: pca_diff sobre respuestas largas captura ruido, no la direccion limpia
- Fix pendiente: dataset formato repeng canonico (mismo prefijo, ultimo token difiere)

NEXT: rework generate_dataset a pares contrastivos cortos estilo repeng emotions example

+++ src/extract_vector.py
@@ -78,7 +78,8 @@ def get_device_and_dtype():
         return "cuda", torch.bfloat16
     if os.environ.get("STEERING_DEVICE") == "mps" and torch.backends.mps.is_available():
         return "mps", torch.float16
-    return "cpu", torch.float32
+    # CPU con fp16 para reducir RAM (fp32 causa OOM en 3.8B)
+    return "cpu", torch.float16
 
 
 def extract(name: str, model_id: str = None, layer_range: tuple[int, int] = None,
